requirements.txt
absl-py==0.7.0
art==3.6
astor==0.7.1
certifi==2019.6.16
coverage==4.5.3
cycler==0.10.0
decorator==4.4.0
editdistance==0.5.2
gast==0.2.2
google-pasta==0.1.7
grpcio==1.18.0
h5py==2.9.0
imageio==2.5.0
imutils==0.5.2
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.8
kiwisolver==1.1.0
lxml==4.3.4
Markdown==3.0.1
matplotlib==3.1.0
mock==3.0.5
networkx==2.3
numpy==1.16.4
opencv-contrib-python==4.1.0.25
opencv-python==4.0.0.21
pandas==0.24.2
Pillow==6.0.0
protobuf==3.6.1
pycm==2.1
pyparsing==2.4.0
python-dateutil==2.8.0
python-docx==0.8.10
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1.1
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
Werkzeug==0.14.1
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
